requirements.txt
google-genai
sqlalchemy
pillow
dotenv
